Only run 2 jobs with leak detection to minimize build times
Motivation:

Running a build with paranoid leak detector level is very expensive and slow. While it is useful to detect buffer leaks it is good enough to just run a few of the builds with it enable and not all. This will still detect leaks while reduce the build time in general a lot.

Modifications:

Change workflow to only run 2 builds with paranoid leak detection enabled.

Result:

Faster build times on the CI