IoUring: Make buffer ring selection more flexible
Motivation:

In the past it was only possible to configure which buffer ring is used per Channel via specify a ChannelOption. This worked but also was a bit unflexible as you might want to use different buffer rings depending on the estimated size of the next read.

Modifications:

- Rename IoUringChannelOption.IO_URING_BUFFER_GROUP_ID to IO_URING_BUFFER_GROUP_ID_HANDLER
- Introduce IoUringBufferRingGroupIdHandler which should be used when you want to use a buffer ring
- Adjust testcases
Result:

More flexible usage of buffer rings possible