Reduce pipeline stack depth (#14782)
Motivation:

This is a forward port of the pipeline stack flattening of #14705.

Modification:

Manual inlining of inbound handler context methods and pipeline methods.

Result:

Flatter stack traces.

There is not much difference in performance. 5.x is already quite a bit
faster than 4.x due to not supporting "child executors".

Some of the difference between 4.x and 5.x also comes down to
implementation differences in `EmbeddedChannel` and it's pipeline.