AdaptivePoolingAllocator: Round chunk sizes up to MIN_CHUNK_SIZE unit… (#14781)
…s and reduce chunk release frequency (#14768)

Motivation:
As the AdaptivePoolingAllocator adjusts it's preferred chunk size, it
may end up causing fragmentation with the system allocator if chunks are
allocated and released frequently.

Modification:
The chunk size is now rounded up to the nearest whole multiple of the
MIN_CHUNK_SIZE. The MIN_CHUNK_SIZE is itself already an even multiple of
many popular small page sizes, like 4k, 16k, and 64k, which makes it
easier for the system allocator to manage the memory in terms of whole
pages.

To further reduce chunk release frequency, chunks are now
probabilistically released if their size differs from the preferred
chunk size.
The probability of a chunk being released depends on how many
MIN_CHUNK_SIZE units they deviate from the preferred chunk size. The
probability is computed as 0.5% per unit of deviation, so chunks that
are off by 2.6 MiB or more are guaranteed to be released.

Result:
Reduced memory fragmentation tendency of the adaptive allocator.