IoUring: Add code to support multishot in general and use ACCEPT_MULTISHOT
Motivation:

Using multishot variants is considered the prefered way to use io_uring to reduce overhead. This change adds the needed changes to support it for "inbound" operations and implement it for accept so far.

Modifications:

- Add code to support multishot for inbound operations in general
- Use ACCEPT_MULTISHOT if possible

Result:

Less overhead for acception connections and preperations to also be able to use multishot for other ops.