More thorough testing of max_cert_list in OPENSSL provider
Motivation:
The existing test was not really covering the functionality, because any configured max certificate list bytes is ignored if it is less than the default 16 KiB.

Modification:
Add more tests for this functionality, particularly ones that set the limit to something greater than 16 KiB, and also add tests that proves the higher limit then accepts certificates that would otherwise fail with the default configuration.
Also adds tests that include session resumption in the mix.

Result:
More thorough testing of the max certificate list bytes setting.