Disable flaky test for now (#14828)
Motivation:
The SignleThreadIoEventLoopTest.testSuspendingWhileRegistrationActive is
flaky.

The event loop and the trySuspend calls are racing, so occasionally a
wakeup call is ignored (due to the internal state), or occasionally the
run(IoHandlerContext) gets to run twice in a row with only one wakeup
call.

When this happens, the event loop misses the suspension signal because
it's waiting on the semaphore, and in turn the currentThread.join() will
not complete.

Modification:
Disable the test for now.

Result:
Build no longer gets stuck on this test.