Fix NPE when upgrade message fails to aggregate (#14816)
Motivation:

When an HTTP message fails to aggregate, for example due to an invalid
'Expect' header, MessageAggregator does not produce a FullHttpRequest.
HttpServerUpgradeHandler would then continue with the next request,
wrongly believing it to be an upgrade request, even though only the
previous one was. This produces an NPE:

```
Caused by: java.lang.NullPointerException: Cannot invoke "java.lang.CharSequence.length()" because "header" is null
	at io.netty.handler.codec.http.HttpServerUpgradeHandler.splitHeader(HttpServerUpgradeHandler.java:429)
	at io.netty.handler.codec.http.HttpServerUpgradeHandler.upgrade(HttpServerUpgradeHandler.java:328)
	at io.netty.handler.codec.http.HttpServerUpgradeHandler.decode(HttpServerUpgradeHandler.java:290)
	at io.netty.handler.codec.http.HttpServerUpgradeHandler.decode(HttpServerUpgradeHandler.java:40)
	at io.netty.handler.codec.MessageToMessageDecoder.channelRead(MessageToMessageDecoder.java:91)
	... 35 common frames omitted
```

Modification:

When the upgrade handler sees a LastHttpContent, cancel the current
upgrade and cancel aggregation.

Result:

No NPE for this input.

Found by fuzzing.