IoUring: Fix IoBufferRing implementation (#14827)
Motivation:

Our implementation made wrong assumptions and so could lead to data
corruption. The problem is that if a ring is exhausted we can't assume
that all our buffers are used already. We need to keep track of what was
used and only refill these when the ring was exhausted.

Motification:

- Fix ring implementation to only refill what was actually used and not
what might be still pending to be used

Result:

Fix buffer ring implementation