IoUring: Make allocations for the buffer ring more flexible (#14838)
Motivation:

At the moment we configure the size of the buffers we want to allocate
to use in buffer ring up front. This is very unflexible as we might want
to either increase or decrease these buffers over the life-time of the
ring.

Modifications:

- Add IoUringBufferRingRecvAllocator that can be used to customize
allocation strategy
- Add IoUringFixedBufferRingRecvAllocator which mimics the old behaviour
by always allocate buffers with the same fixed size
- Change IoUringBufferRingConfig to take
IoUringFixedBufferRingRecvAllocator as argument and remove chunkSize and
allocator

Result:

More flexible when allocating buffers for the buffer ring. Fixes
https://github.com/netty/netty/issues/14805