Add QueryStringDecoder option to leave '+' alone Motivation:
The URI standard RFC 3986 does not specify that query components have their spaces encoded as `+`. It is implied that the encoding is `%20` instead. However, the whatwg HTML standard says explicitly that the query must be encoded using `application/x-www-form-urlencoded` rules, which does use `+` for space. This is also what browsers do.

QueryStringDecoder should offer a way to parse either format.

Modification:

- Modify QueryStringDecoder to use a builder to accommodate the increasing number of flags
- Add a `htmlQueryDecoding` flag, enabled by default, to control the `+` decoding

The default value of `htmlQueryDecoding` is appropriate for most use cases, I don't think it should be changed even in netty 5.

Result:

Query strings encoded purely according to the URI spec can be decoded properly.