Convenient API for specifying SNI hostname (#14849)
Motivation:
As we now depend on Java 8, we can now mention these new types in our
API. This allow us to let clients configure their TLS SNI extension
through the `SslContextBuilder`.

Modification:
Add a method to the `SslContextBuilder` for adding SNI server names
extensions. Funnel the parameters through to the relevant `SSLEngine`
implementations and their `SSLParameter` instances. Update the test that
was testing the SNI functionality to use the new APIs.

Result:
We now have a more convenient API for specifying SNI server names in the
TLS `ClientHello` message.

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>