IoUring: Correctly handle pollIn notifications if there is no read pending
Motivation:

It is possible that while a pollIn notification is received we don't have a read pending yet. This is especially true when multishot polladd is used.

Modification:

Only try to schedule a read if a read is pending and if not set the correct internal state that once read() is triggered we know that there is something to read.

Result:

Correctly handle pollIn notifications in all cases