IoUring: Use IORING_SETUP_CLAMP when setup the ring
Motivation:

There is no way to know what the maximum size is that is supported by the kernel. Let's use IORING_SETUP_CLAMP to allow the kernel to clamp it if needed.

Modification:

Use IORING_SETUP_CLAMP (which is supported till kernel 5.6)

Result:

Don't fail if a ring size is used that is not supported by the kernel