IoUring: Correctly handle late removal of registration for multishot completions
Motivation:

We keep track of outstanding completions to correctly handle the removal of the registration. The problem was that we assumed that one submission always restults in one completion. This is not always true and so we sometimes removed a registration even tho there were more completions pending.

This resulted in logs like:

```
2025-03-10T02:51:25.6273390Z 02:51:25.614 [testsuite-io_uring-worker-19-3] DEBUG io.netty.channel.uring.IoUringIoHandler -- ignoring POLL_ADD completion for unknown registration (id=-2147483624, res=-125)
```
Modifications.

Before decrementing the outstandingCompletions counter check if the IORING_CQE_F_MORE flag is not set

Result:

Correctly handle  late removal