IoUring: Use correct native jni types in function definition (#14914)
Motivation:

When define functions that are exposed via JNI we need to ensure we use
the types that are provided and supported by JNI. Not doing so might
result in undefined behaviour depending on the platform.

Modification:

Fix netty_io_uring_unregister_buf_ring to use the correct supported
types.

Result:

No undefined behaviour anymore when unregister buffer ring