IoUring: Only cancel operations once
Motivation:

cancelOps(...) might be called multiple times while trying to teardown a channel. There is not need to submit multiple ASYNC_CANCEL for the same operation, so guard against it.

Modifications:

Just set the ids to 0 after the cancel was submitted and guard against submitting a cancel if the id is 0

Result:

Less submissions during teardown of a Channel