IoUring: Fix AssertionError caused by racing cancellation and completion (#14921)
Motivation:

There is a possibility of having the cancellation race with processing
the completion from a previous operation. Like it might already be
completed but just not processed yet by us. Because of this we can't
assert that the id is 0

Modifications:

- Remove asserts and simplify

Result:

Fixes https://github.com/netty/netty/issues/14915