IoUring: Fix crashes caused by IoUringBufferRing
Motivation:

Our current implementation of IoUringBufferRing sometimes cause a crash which seems to be caused by the fact that we use a sub-type of WrappedByteBuf. While we investigate why this happens let's revert the change that did introduce this crash for now.

Modifications:

Remove max unreleasable buffer tracking code for now.

Result:

No more crashes