IoUring: Throw IllegalArgumentException if IOSQE_CQE_SKIP_SUCCESS is … (#14927)
…used

Motivation:

Because of how we keep track of submissions and completions its
impossible to support IOSQE_CQE_SKIP_SUCCESS for now. We should guard
against using of it as it would cause problems when we try to keep track
on when we can finally remove a registration.

Modifications:

- Check if IOSQE_CQE_SKIP_SUCCESS is used and if so throw an
IllegalArgumentException
- Add unit test

Result:

Guard against usage of unsupported flag