IoUring: Ensure close completes before closing the eventloop group (#14933)
Motivation:

Currently we have a race in our implementation which can cause a segfault when the event loop is shutdown before the channel is closed completely. This sometimes causes failures on the CI. While we investigate this race let's ensure we correctly close channels before teardown the eventloop to make the build stable again.

Modification:

- Block until Channel is really closed.

Result:

No more segfaults.