Disable sun.misc.Unsafe by default on Java 24+
Motivation:
Java 24 has started to print warnings when `sub.misc.Unsafe` is used to access memory.
See JEP 498: https://openjdk.org/jeps/498

People tend to be sensitive to new warnings and errors printed by the JVM, so we should avoid enabling `Unsafe` by default on Java 24 onwards.

Modification:
Effectively change the default value of the `io.netty.noUnsafe` system property for Java 24 onwards, but in a way where we can tell if it was explicitly set to a given value or if it follows the default.

Any sort of explicit settings, be it `-Dio.netty.noUnsafe=false` or `--sun-misc-unsafe-memory-access=allow`, will continue to work as before.

Manually verified that the changes work as intended on Java 24.

Result:
No more warnings printed to the console on Java 24+.

Fixes https://github.com/netty/netty/issues/14942