Merge commit from fork
* Use SipHash to hash connection id to prevent form hash colision attack.

Motivation:

We need to use a cryptographic hashing algo to hash as otherwise it might be possible by the remote peer to cause hashing colisions and so consume a lot of CPU on the server.

See https://github.com/ncc-pbottine/QUIC-Hash-Dos-Advisory

Modifications:

Implement our own HashMap that uses SipHash to do the mapping for connection id and connection.

Result:

Not possible for the remote peer to cause colision attack

* Fix implementation and add tests from paper

* Use hashmap

* Implement Comparable

* Address comment

* Update codec-classes-quic/src/main/java/io/netty/incubator/codec/quic/ConnectionIdChannelMap.java

Co-authored-by: Bryce Anderson <bl_anderson@apple.com>

* Update codec-classes-quic/src/main/java/io/netty/incubator/codec/quic/ConnectionIdChannelMap.java

Co-authored-by: Jonas Konrad <me@yawk.at>

* Remove cast

---------

Co-authored-by: Bryce Anderson <bl_anderson@apple.com>
Co-authored-by: Jonas Konrad <me@yawk.at>