Remove dummy head entry in DefaultHeaders.
Motivation:

DefaultHeaders relies on a dummy head entry instance to maintain the linked list of headers. While this is convenient, this is not strictly necessary.

Changes:

Remove the dummy head entry and assume that the default head is null.

When an entry is added, it becomes the new head if the head was previously null. Likewise if the last entry is removed, the head becomes null again.

Book keeping is required to maintain the new invariant.

Iteration checks the nullity of the head before.

Special care also is done in DefaultHttp2Headers implementation which interacts with the protected state of the parent DefaultHeaders.

Results:

This saves instantiating a dummy head element per headers entry.

It also breaks subclasses using the leaky DefaultHeaders fields (which might veto this change).