Allow delayed set of owningThread in ManualIoEventLoop Motivation:
In some scenarios, it is desirable to immediately create a ManualIoEventLoop, without knowing which thread it is assigned to. In particular, to implement `MultiThreadIoEventLoopGroup.newChild` properly, the owning thread is assigned through an `Executor` call, so we don't see it until the `Executor` actually invokes the `Runnable` it receives, but we need to return the `ManualIoEventLoop` from `newChild` immediately.

Modification:

Allow passing `null` for the `owningThread` constructor parameter, and add a `setOwningThread` setter that can be called at most once.

Result:

`newChild` can be implemented without hacks.

The only use of the `owningThread` field is in `inEventLoop`. That means that the only difference compared to the old constructor approach is that there is a period of time when the future owning thread is running, but inEventLoop still returns `false`. I'm struggling to imagine scenarios where this matters, but @franz1981 had some thoughts, so I'm open to concerns and alternative solutions.