Fix support for DH_anon and ECDH_anon authentication methods
Motivation:
We have tests for these methods, and the tests aren't running because the OpenSSL version we test with is too old.
However, when we run the tests on CentOS 7, they start running and they fail because our server-side cert callback tries to choose key material for these auth methods, decides it can't, and throws an exception that breaks the handshake.

The DH_anon and ECDH_anon auth methods are only relevant when we hook up a callback in tcnative to provide ephemeral DH key material to openssl, but this is not enabled by default.

Modification:
- Add an `OpenSslContextOption.TMP_DH_KEYLENGTH` so we can set DH key material on a per-test basis.
- Update the relevant tests to enable DH ephemeral keys.
- Update the key material callbacks to allow these authentication methods to go through, when ephemeral DH keys are enabled.

Result:
The DH_anon and ECDH_anon authentication methods are no longer broken.
The tests should now pass when running on a new enough OS.