Build Netty 4.2 on CentOS 7 (#14973)
Motivation:
This relatively small version bump will help us with graduating the QUIC
incubator module.

Modification:
Add a CentOS 7 base docker image, and convert all 4.2 builds to use it
(except for the cross-compile builds that have separate images).

The converted builds include the PR builds, the branch builds, the
deploy builds, and the 4.2 release build.

The CentOS 6 image is still staying because it is used by the 4.1
release workflow which is defined in both 4.1 and 4.2 branches.

Result:
Netty 4.2 is now built on CentOS 7 instead of CentOS 6. This will help
us with graduating the QUIC module, because it's builds are already all
on CentOS 7.