Add suppressed exception to original cause when calling Future.sync* (#14985)
Motivation:
The `Promise.sync` and `syncUninterruptibly` methods will rethrow any
exception that cause the promise to fail. This gives the stack trace of
the original failure, but people loose the stack trace telling which
specific sync call propagated the exception. This can make debugging
confusing and more difficult that necessary. We have to keep this
behavior by default for compatibility, but we can add the extra details
as suppressed exception and so help with debugging.

Modification:
Add`CompletionException` when `sync` or `syncUninterruptibly` is called
as suppressed exception to the original cause.

Result:
Easier to debug.

This is a. back port of https://github.com/netty/netty/pull/14898

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>