Fix timeout when auto read is disabled late in io_uring Motivation:
If auto read is disabled late (during a channelRead), but there is a read call during the next channelReadComplete, that read call would be ignored, leading to the channel freezing up and the connection timing out.

Modification:

When clearRead cancels the scheduled POLLIN, clear the iostate immediately, so that the next `read()` will correctly schedule a new POLLIN.

Not sure if this is the proper way to do it. Maybe this should go into cancelComplete0 instead? But then we'd have to add some mechanism to hold back the read call. I think the risk of duplicate reads that this change adds is better than no reads.

Result:

Echo server does not freeze up anymore after turning off auto read during a read operation.