[bugfix]: Fix the issue where pendingTimeouts may be incorrect in the… (#15000)
… HashedWheelTimer. (#14959) (#14960)

Motivation:

If the user sets a positive `maxPendingTimeouts` of `HashedWheelTimer,`
this bug means that the maximum number of tasks can be potentially
increased.

Modification:

When the worker thread expires the tasks of the current tick,
`pendingTimeouts` decremented only when the status of
`HashedWheelTimeout` is successfully changed from `ST_INIT` to
`ST_EXPIRED.` For other cases, `pendingTimeouts` will be decremented in
`processCancelledTasks()` when the next tick is reached.

Result:

Fixes #14959.

Co-authored-by: Walter Jia <walterjia@foxmail.com>