fix: WebSocketClientCompressionHandler shouldn't claim window bits support when jzlib is not availablets value is not supported
Motivation:

WebSocketClientCompressionHandler and PerMessageDeflateClientExtensionHandshaker use ZlibCodecFactory.isSupportingWindowSizeAndMemLevel(), which in netty 4.2 always returns true.

Actually, `java.util.zip.Deflater` doesn't support window bits and only works if the server provides a value of 15, which is unlikely and should only be considered a lucky optimization path.

Different values would only work if `jzlib` is available in the classpath. But it's an optional dependency and it's less performant than the JDK implementation as it's pure Java.

Currently, when jzlib is not available in the classpath and the server decides to take over client window bits with a value different from 15 (eg python would use 12), Netty crashes with a CNFE.

Motivation:

Only advertise window bits support if jzlib is available in the classpath.

Result:

No more CNFE when the WebSocket server reacts to client bits support with a non 15 value and  jzlib is not available in the classpath.