SSL: Clear native error if named group is not supported (#15001) (#15038)
Motivation:

We use SSLContext.setCurvesList(...) to detect if a named group is not
supported or not but if its not we don't clear the native error stack
atm. This is problematic as if you miss to clear the stack it might
produce errors later which might cause the native SSL context creation
to fail.

Modifications:

Add SSL.clearError()

Result:

Fixes https://github.com/netty/netty/issues/14991