Add ticker to ManualIoEventLoop (#15043)
Motivation:

ManualIoEventLoop is useful for testing because it matches "normal"
event loops more closely than EmbeddedEventLoop. Adding ticker
functionality makes testing even easier.

Modification:

Add a ticker parameter.

Note that the IoHandler *does not* use the ticker. A blocking `run`
operation may block for the specified real-time duration even if the
ticker time flows differently. I don't believe this can be fixed: The
IoHandler uses system APIs to block, so the only way to make it adhere
to the Ticker would be to wake the IoHandler up when the Ticker time
changes, but this would require the Ticker to keep track of the
IoHandlers that use it. I don't think this is a good idea.

Since ManualIoEventLoop has a wakeup method, this is easy to work around
for users.

Result:

Easy testing of time-dependent code with ManualIoEventLoop.