IoUring: Ensure that the buffer group is properly set before channel::read (#15054)
Motivation:

Due to an issue with the promise callback sequence in the `doRegister` ,
even if the BufferGroup is set, the first recv of the SocketChannel will
not use the buffer group.


![error](https://github.com/user-attachments/assets/d25e3d27-e1e0-4e0e-b8cb-2976ccde7f62)

Therefore, we need to adjust the callback sequence—call
ioUringIoHandler.findBufferRing first, and then call the beginRead.

Modification:

Ensure that the buffer group is properly set before channel::read

Result:
The fixed result is:

![right](https://github.com/user-attachments/assets/a3e30f98-b20d-4d25-a22e-6ed097de8725)