Fix transport-native-io_uring Bundle-SymbolicNames
Motivation:

OSGi expects the combination of 'Bundle-SymblicName' and
'Bundle-Version' to be unique to a particular bundle, but we generate a
bundle for each jniClassifier.

Modification:

Add a bundle-plugin invocation for each of the native profiles,
appending {$jniClassifier} to what would normally be generated.

Result:

Multiple CPU architectures can be supported by a single Apache Karaf feature.

Signed-off-by: Robert Varga <robert.varga@pantheon.tech>