Fix transport-native-epoll Bundle-SymbolicNames (#15059) (#15068)
Motivation:

OSGi expects the combination of 'Bundle-SymblicName' and
'Bundle-Version' to be unique to a particular bundle, but we generate a
bundle for each jniClassifier.

Modification:

Add a bundle-plugin invocation for each of the native profiles,
appending {$jniClassifier} to what would normally be generated. This
also provide an anchor for OSGi-specific manifest entries, so we move
Bundle-NativeCode and Fragment-Host declaration.

Result:

Multiple CPU architectures can be supported by a single Apache Karaf
feature.

Motivation:

Explain here the context, and why you're making that change.
What is the problem you're trying to solve.

Modification:

Describe the modifications you've done.

Result:

Fixes #15059. 

Signed-off-by: Robert Varga <robert.varga@pantheon.tech>