Fix reserved key confusion caused by unchecked cast (#328)
Motivation:
In the method `Http3CodecUtils.isReservedHttp2Setting` the long key
parameter is user-supplied data, so we can't blindly cast it to an int.
It may contain higher-order bits which would be truncated off, and thus
leading us to confuse the value with a reserved key.

Modification:
Remove the int-cast and compare the value directly with literal
long-values.

Result:
Avoid confusion from a truncated int cast of a user-supplied value.