Fix init order of PlatformDependent0 fields (#15077)
Motivation:

After #14975, `EXPLICIT_NO_UNSAFE_CAUSE` has a dependency to
`JAVA_VERSION`, which is initialized after it. When
`explicitNoUnsafeCause0()` tries to set `io.netty.noUnsafe` to true by
default on Java 24, it fails, because `javaVersion()` always returns 0
at that point.

Also, `isAndroid0()` is called twice, once in `javaVersion0()` and once
to initialize `IS_ANDROID`. This is not a bug or a performance issue,
but it does have a visible side-effect when debug logging is enabled.

Modification:

The declaration of `EXPLICIT_NO_UNSAFE_CAUSE` is moved after
`JAVA_VERSION`, ensuring `javaVersion()` returns the correct version
when `explicitNoUnsafeCause0()` calls it.

`IS_ANDROID` is also moved before `JAVA_VERSION`, such that the
`isAndroid0()` method is called only once.

Result:

Fixes #14942 correctly