Optimize capacity bumping for adaptive ByteBufs
Motivation:
It is quite common in downstream code to not specify a buffer size when allocating.
People instead rely on the ByteBuf ability to increase its size automatically.
The ByteBufs allocated by the adaptive allocator also support this, but has thus far required a round-trip through the allocator every time.
In contrast, the pooled allocator is reserving the chunk run-size when allocating buffers, and very cheaply allow smaller buffers to bump up to the run-size limit.
This is an important performance optimisation in practice.
We can do something similar in the adaptive allocator, because we keep statistics about the buffer sizes that we allocate.

Modification:
The adaptive allocator now reserves space equal to the 99-percentile buffer size, when allocating buffers, and allow its ByteBufs to cheaply bump their capacities up to this limit.
We already compute this percentile during histogram rotation, so it is simply a matter of storing this value in the magazine, and then using it when reserving space out of chunks.

Result:
Calls like ByteBuf.ensureWritable() will now on average be much cheaper when using the adaptive allocator.