Cleanup code to make stacktrace less confusing (#796)
Motivation:

5dade1f2b87504f8e4e2b385f62d63f537aac244 introduced a change that did
prevent the Channel to be closed if a STREAM_STOPPED frame was received
but did not change the name of the method from forceClose(...) to
something else. This makes the stacktrace quite confusing as while the
method is called forceClose(...) it actually might not close the
channel.

Modifications:

- Move all the logic into the writable(...) method and so make things
less confusing

Result:

Less confusing stacktrace for failed writes