Revert "Try to give our docker images more disk storage (#15071)" (#15101)
Motivation:

The added storage option is not supported on the fs used by gh actions,
causing an error to be printed and the job aborts:
https://github.com/netty/netty/actions/runs/14710201216/job/41280392533

```
Error response from daemon: --storage-opt is supported only for overlay over xfs with 'pquota' mount option
```

Modifications:

Revert commit 592e05bc746959f0af98c2ecdb142119c257fb8b which added the
option that is causing the failure. Same was done for 4.2 in #15074.

Result:

Action should not fail anymore due to the unsupported option.